+++ test/fixture/applications/testPassportLocal.js
