--- test/fixture/login_local/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "carcass-auth-fixture-local",
-  "version": "0.0.0",
-  "dependencies": {}
-}
