--- test/fixture/login_session/index.js
@@ -1,6 +1,0 @@
-var carcass = require('carcass');
-
-require('../../..');
-
-// Register applications.
-carcass.register(__dirname, 'applications');
