+++ test/fixture/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "carcass-auth-fixture-session",
+  "name": "carcass-auth-fixture",
   "version": "0.0.0",
   "dependencies": {}
 }
